factor GetWtNames into smaller functions
Refactor the GetWtNames function, extracting the logic to smaller helper functions.

Introduce a new GetWtDutyList function which returns a list of the duty names instead of keeping them in a single string. It also reports the number of needed objects via an output parameter.

Refactor the duty name reduction to a single function which takes a duty name and attempts to reduce it. Call this via Select() to translate all the duty names in the duty list.

Replace the home-grown method of joining strings with string.join, and refactor the functions to pass the separator string instead of passing a "copy" or "list" value.

Since we no longer directly call GetWTNames with forced parameters, simplify the code to avoid using that parameter entirely. If we ever need this in the future, we can make use of optional booleans and the ??= operator.