factor out GetWTDutyList() function from GetWTNames()
The GetWTNames() function determines the list of duty names as a string
by checking the wondrous tails book and finding the named values from
the excell sheets.

Extract this sheets logic to its own function. This function returns the
duty names as a list of strings, instead of a single pre-joined string.
It also returns the number of remaining objectives as the output
variable numNeeded.

Doing this allows limiting the amount of code that needs to be wrapped
in unsafe, and makes it easier to refactor other aspects of the
GetWTNames() function in a following commit.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>