factor duty name reduction to ReduceWTDutyName()
The GetWTNames() function has a huge block of code which is used to
shorten the names of wondrous tails duties by some custom regex rules.

Extract this logic to a separate function, so that we can later use it
in a Select() call. This also helps reduce the amount of indentation,
and keep functions smaller and easier to process.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>