convert reducedText to a Select call
The reducedDuty configuration option is implemented by calling the
ReduceWTDutyName function on each duty in the dutyList as part of the
loop which joins the names together.

A future change is going to remove this loop in favor of using
string.join.

Rather than keeping the loop, notice that this logic can be thought of
as a mapping operation where we map from the standard name to its
reduced name. In C# this is known as the Select() operation, where
Select invokes a function on each element of the collection, returning a
new collection.

Convert the reducedTest code to use this Select notation.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>