add "need" string to dutyList before joining strings
Add the final "need <N>" string to the dutyList before joining the
strings together in a loop. I separated this out since it makes the
resulting diff easier to follow.

Note that this does have one change which is that we now add the need
line for all modes of this function rather than just the "copy" mode. I
think thats acceptable, but this might need to be adjusted.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>