use GetWTDutyList in RemainingObjections()
The RemainingObjectives() function needs to obtain the number of
objectives remaining. It currently does this by calling GetWtNames(),
with a force parameter telling the function to override configuration
and list the needed value last. Then, it extracts this from the string.

This can be simplified to get the value directly from GetWTDutyList()
without any string munging. This also allows removing the force options
from the GetWTNames() as no caller needs these any more.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>